removes src. fixes bans new tanks bugfixes etc

+++ Interstation-Two-WW2/code/modules/admin/verbs/quickBan.dm
@@ -2,18 +2,42 @@
  * not nearly as fancy as old banning, and it wasn't worth making an interface
  * for this rather small amount of code, so its all done via BYOND's input() */
 
-/* this is a WIP, currently only server bans are supported */
+// 7 ways to ban people
 var/list/ban_types = list("Job Ban", "Faction Ban", "Officer Ban", "Server Ban", "Observe Ban", "Playing Ban")
+
+/datum/quickBan_handler
+/datum/quickBan_handler/Topic(href,href_list[])
+	..()
+	if (href_list["quickBan_removeBan"])
+		var/client/caller = locate(href_list["caller"])
+		var/UID = href_list["quickBan_removeBan_UID"]
+		var/ckey = href_list["quickBan_removeBan_ckey"]
+		var/cID = href_list["quickBan_removeBan_cID"]
+		var/ip = href_list["quickBan_removeBan_ip"]
+		database.execute("DELETE FROM quick_bans WHERE (UID = '[UID]' OR ckey = '[ckey]' OR cID = '[cID]' OR ip = '[ip]');")
+		var/M = "[key_name(caller)] removed a ban for '[UID]/[ckey]/[cID]/[ip]'."
+		log_admin(M)
+		message_admins(M)
+
+		for (var/client/C in clients)
+			if (C.ckey == ckey)
+				C << "<span class = 'good'>href_list["quickBan_unbanned_message"]</span>"
+
+
+var/datum/quickBan_handler/quickBan_handler = null
 
 /* admin procedures */
 /client/proc/quickBan_search()
 	set category = "Bans"
 
+	if (!quickBan_handler)
+		quickBan_handler = new
+
 	var/option = input(src, "Search for a ban?") in list("Yes", "No")
 	if (option == "No")
 		return
 
-	var/_ckey = input(src, "What ckey will you search for? (optional)") as null|text
+	var/_ckey = ckey(input(src, "What ckey will you search for? (optional)") as null|text)
 	var/cID = input(src, "What cID will you search for? (optional)") as null|text
 	var/ip = input(src, "What address will you search for? (optional)") as null|text
 	var/ban_type = input(src, "What type of ban do you want to search for?") in ban_types + "All"
@@ -25,7 +49,8 @@ var/list/ban_types = list("Job Ban", "Faction Ban", "Officer Ban", "Server Ban",
 	if (ban_type == "All")
 		result = database.execute("SELECT * FROM quick_bans;", FALSE)
 	else
-		result = database.execute("SELECT * FROM quick_bans WHERE ban_type == '[ban_type]';", FALSE)
+		ban_type = replacetext(ban_type, " Ban", "")
+		result = database.execute("SELECT * FROM quick_bans WHERE type = '[ban_type]';", FALSE)
 
 	var/list/possibilities = list()
 
@@ -38,9 +63,10 @@ var/list/ban_types = list("Job Ban", "Faction Ban", "Officer Ban", "Server Ban",
 			if (ip && result.Find("ip_[v]") && result["ip_[v]"] != ip)
 				continue
 			if (text2num(result["expire_realtime_[v]"]) <= world.realtime)
-				database.execute("REMOVE * FROM quick_bans WHERE UID = '[result["UID_v"]]';")
+				database.execute("DELETE * FROM quick_bans WHERE UID = '[result["UID_v"]]';")
 				continue
-			possibilities += "<big><b>UID [result["UID_[v]"]]</b> (<a href='byond://?src=\ref[src];quickBan_removeBan=1;quickBan_removeBan_UID=[result["UID_[v]"]];quickBan_removeBan_ckey=[result["ckey_[v]"]];quickBan_removeBan_cID=[result["cID_[v]"]];quickBan_removeBan_ip=[result["ip_[v]"]]'>DELETE</a>)</big>: [result["ckey_[v]"]]/[result["cID_[v]"]]/[result["ip_[v]"]], type '[result["type_[v]"]]' ([result["type_specific_info_[v]"]]): banned for '[result["reason_[v]"]]' by [result["banned_by_[v]"]] on [result["ban_date_[v]"]]. <b>[result["expire_info_[v]"]]</b>. (After assigned date)"
+			var/quickBan_unbanned_message = "Your ban [result["ban_type"]] ([result["type_specific_info"]]) has been lifted by an admin."
+			possibilities += "<big><b>UID [result["UID_[v]"]]</b> (<a href='byond://?src=\ref[quickBan_handler];caller=\ref[src];quickBan_removeBan=1;quickBan_removeBan_UID=[result["UID_[v]"]];quickBan_removeBan_ckey=[result["ckey_[v]"]];quickBan_removeBan_cID=[result["cID_[v]"]];quickBan_removeBan_ip=[result["ip_[v]"]];quickBan_unbanned_message=[quickBan_unbanned_message]'>DELETE</a>)</big>: [result["ckey_[v]"]]/[result["cID_[v]"]]/[result["ip_[v]"]], type '[result["type_[v]"]]' ([result["type_specific_info_[v]"]]): banned for '[result["reason_[v]"]]' by [result["banned_by_[v]"]] on [result["ban_date_[v]"]]. <b>[result["expire_info_[v]"]]</b>. (After assigned date)"
 
 	for (var/possibility in possibilities)
 		html += "<br>"
@@ -61,6 +87,8 @@ var/list/ban_types = list("Job Ban", "Faction Ban", "Officer Ban", "Server Ban",
 		fields["ckey"] = input(src, "What is the person's ckey? (optional)") as null|text
 		fields["cID"] = input(src, "What is the person's cID? (optional)") as null|text
 		fields["ip"] = input(src, "What is the person's IP? (optional)") as null|text
+		if (!fields["ckey"] && !fields["cID"] && !fields["ip"])
+			return
 	else if (option == "Client")
 		var/client/C = input(src, "Which client?") in clients + "Cancel"
 		if (C == "Cancel")
@@ -127,12 +155,16 @@ var/list/ban_types = list("Job Ban", "Faction Ban", "Officer Ban", "Server Ban",
 		if (29.99 to 364.99) // count in months
 			fields["expire_info"] = "Expires in [duration_in_days/30] month(s)"
 		if (364.99 to INFINITY) // count in years
-			fields["expire_info"] = "Expires in [duration_in_days/365] years(ss"
+			fields["expire_info"] = "Expires in [duration_in_days/365] year(s)"
 
 	if (global_game_schedule)
 		fields["ban_date"] = global_game_schedule.getDateInfoAsString()
 
+	reenter_reason
 	fields["reason"] = input(src, "Provide a reason for the ban.") as text
+	if (!fields["reason"])
+		goto reenter_reason
+
 	fields["banned_by"] = key
 
 	quickBan_ban(fields, src)
@@ -140,11 +172,11 @@ var/list/ban_types = list("Job Ban", "Faction Ban", "Officer Ban", "Server Ban",
 /* helpers */
 /proc/quickBan_sanitize_fields(var/list/fields)
 
-	if (!fields.Find("ckey"))
+	if (!fields.Find("ckey") || fields["ckey"] == "")
 		fields["ckey"] = "nil"
-	if (!fields.Find("cID"))
+	if (!fields.Find("cID") || fields["cID"] == "")
 		fields["cID"] = "nil"
-	if (!fields.Find("ip") || fields["i"] == null) // host
+	if (!fields.Find("ip") || fields["ip"] == "" || fields["ip"] == null) // host
 		fields["ip"] = "nil"
 	if (!fields.Find("type"))
 		fields["type"] = "nil"
@@ -205,7 +237,7 @@ var/list/ban_types = list("Job Ban", "Faction Ban", "Officer Ban", "Server Ban",
 	if (database.execute("INSERT INTO quick_bans (ckey, cID, ip, type, type_specific_info, UID, reason, banned_by, ban_date, expire_realtime, expire_info) VALUES ('[fields["ckey"]]', '[fields["cID"]]', '[fields["ip"]]', '[fields["type"]]', '[fields["type_specific_info"]]', '[fields["UID"]]', '[fields["reason"]]', '[fields["banned_by"]]', '[fields["ban_date"]]', '[fields["expire_realtime"]]', '[fields["expire_info"]]');"))
 		if (banner)
 			banner << "<span class = 'notice'>You have successfully banned [ckey]/[cID]/[ip]. This ban [lowertext(expire_info)]."
-		var/M = "[key_name(banner)] banned [ckey]/[cID]/[ip] (bantype = [fields["type"]]) for reason '[fields["reason"]]'. This ban [lowertext(expire_info)]."
+		var/M = "[key_name(banner)] banned [ckey]/[cID]/[ip] (bantype = [fields["type"]] ([fields["type_specific_info"]])) for reason '[fields["reason"]]'. This ban [lowertext(expire_info)]."
 		log_admin(M)
 		message_admins(M)
 		// kick whoever got banned if they're on
@@ -218,7 +250,7 @@ var/list/ban_types = list("Job Ban", "Faction Ban", "Officer Ban", "Server Ban",
 			if (fields["type_specific_info"])
 				for (var/client/C in clients)
 					if (C.ckey == ckey)
-						C << "<span class = 'userdanger'>You have been [fields["type"]]-banned ([fields["type_specific_info"]]). Reason: '[fields["reason"]]'. This ban [lowertext(expire_info)]."
+						C << "<span class = 'userdanger'>You have been [lowertext(fields["type"])]-banned ([fields["type_specific_info"]]). Reason: '[fields["reason"]]'. This ban [lowertext(expire_info)]."
 						break
 			else
 				for (var/client/C in clients)
@@ -229,30 +261,50 @@ var/list/ban_types = list("Job Ban", "Faction Ban", "Officer Ban", "Server Ban",
 		if (banner)
 			banner << "<span class = 'warning'>FAILED to ban [ckey]/[cID]/[ip]! A database error occured.</span>"
 
-/* checking if we're banned & then reject us */
+/* checking if we're banned */
 /client/proc/quickBan_isbanned(var/ban_type = "Server", var/type_specific_info = "nil")
-	var/list/bans = database.execute("SELECT * FROM quick_bans WHERE (ckey = '[ckey]' OR cID = '[computer_id]' OR ip = '[address]') AND type = '[ban_type]';", FALSE)
+	var/list/bans = database.execute("SELECT * FROM quick_bans WHERE (ckey = '[ckey]' OR cID = '[computer_id]' OR ip = '[address]') AND type = '[ban_type]' AND type_specific_info = '[type_specific_info]';", FALSE)
 	if (islist(bans) && !isemptylist(bans))
 		for (var/x in bans)
 		//	world << "[x] = [bans[x]]"
 			if (x == "expire_realtime" && text2num(bans[x]) <= world.realtime)
-				database.execute("DELETE FROM quick_bans WHERE UID = '[bans["UID"]]';")
+				if (bans.Find("UID"))
+					database.execute("DELETE FROM quick_bans WHERE UID = '[bans["UID"]]';")
 				continue
-			if (x == "reason")
+			else if (x == "reason")
 				if (bans.Find("expire_realtime") && text2num(bans["expire_realtime"]) <= world.realtime)
 					database.execute("DELETE FROM quick_bans WHERE UID = '[bans["UID"]]';")
 					continue
 				if (bans.Find("type_specific_info"))
-					if (bans["type_specific_info"] == type_specific_info)
-						return bans[x]
-				return FALSE
+					log_debug(bans["type_specific_info"])
+					log_debug(type_specific_info)
+					if (bans["type_specific_info"] != "nil")
+						if (bans["type_specific_info"] == type_specific_info)
+							return list("reason" = bans["reason"],
+								"ban_date" = bans["ban_date"],
+								"expire_info" = bans["expire_info"])
+					else
+						return list("reason" = bans["reason"],
+							"ban_date" = bans["ban_date"],
+							"expire_info" = bans["expire_info"])
+				else
+					return list("reason" = bans["reason"],
+						"ban_date" = bans["ban_date"],
+						"expire_info" = bans["expire_info"])
 	return FALSE
 
 /* check if we're banned and tell us why we're banned */
 /client/proc/quickBan_rejected(var/bantype = "Server")
-	var/banreason = quickBan_isbanned(bantype)
-	if (banreason)
-		src << "<span class = 'userdanger'>You're [lowertext(bantype)]-banned. Reason: '[banreason]'</span>"
+	var/list/fields = quickBan_isbanned(bantype)
+	var/reason = fields["reason"]
+	var/date = fields["ban_date"]
+	var/expire_info = fields["expire_info"]
+
+	if (reason)
+		if (bantype == "Server")
+			src << "<span class = 'userdanger'>You're banned. Reason: '[reason]'. This ban was assigned on [date] and [expire_info] (after assigned date)</span>"
+		else
+			src << "<span class = 'userdanger'>You're [lowertext(bantype)]-banned. Reason: '[reason]'. This ban was assigned on [date] and [expire_info] (after assigned date)</span>"
 		return TRUE
 	return FALSE
